+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 0.0.11 / 2014-08-24
+* [FIX] `SequiturGrammar#check_unicity`: an exception was raised when it shouldn't. Added example in spec file.
+* [CHANGE] `sequitur.rb` : Added the convenience Sequitur::build_from method.
+
 ### 0.0.10 / 2014-08-24
 * [CHANGE] `README.md`: Added hyperlinks about Sequitur algorithm.